chore(back): Split JSON and JSONL exporters
