+++ pixano/datasets/exporters/__init__.py
@@ -5,7 +5,12 @@
 # =====================================
 
 from .dataset_exporter import DatasetExporter
-from .default import DefaultJSONDatasetExporter, DefaultJSONLDatasetExporter
+from .default_json_dataset_exporter import DefaultJSONDatasetExporter
+from .default_jsonl_dataset_exporter import DefaultJSONLDatasetExporter
 
 
-__all__ = ["DatasetExporter", "DefaultJSONDatasetExporter", "DefaultJSONLDatasetExporter"]
+__all__ = [
+    "DatasetExporter",
+    "DefaultJSONDatasetExporter",
+    "DefaultJSONLDatasetExporter",
+]
